refactor : types 폴더 내에 있는 파일들을 이동
- common 폴더에서 types 폴더로 이동
